// apps/web/src/api/model/sourceSearchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Scholia API
 * OpenAPI spec version: 0.1.0
 */
import type { SourcePersonResponse } from "./sourcePersonResponse.ts";
import type { SourceType } from "./sourceType.ts";

export interface SourceSearchResponse {
    created_by: string;
    id: string;
    persons: SourcePersonResponse[];
    protected: boolean;
    /** @nullable */
    publication_year?: number | null;
    source_type: SourceType;
    title: string;
}